+++ ingest2/main.go
@@ -4,7 +4,7 @@
 package ingest2
 
 // Add new version and assign it to `CurrentIngestVersion` if you want force reingest (after backward not compatible changes)
-const CurrentIngestVersion = IngestVersionUnmatchedSaleParticipation
+const CurrentIngestVersion = IngestVersionUnlockedEffectAfterLocked
 
 const (
 	IngestVersionInitial = iota
@@ -14,4 +14,5 @@ const (
 	IngestVersionRecoveryState
 	IngestVersionKYCRecoveryAutoApprove
 	IngestVersionUnmatchedSaleParticipation
+	IngestVersionUnlockedEffectAfterLocked
 )